server/requirements.txt
Django==1.9
djangorestframework==3.3.2
mysqlclient==1.3.4  # Python 3 - Windows compatible version
django-cors-headers==1.1.0
git+git://github.com/CptLemming/django-json-patch.git#egg=django-json-patch
